backend/requirements.txt
# Core dependencies
fastapi==0.110.0
uvicorn[standard]==0.29.0

# AI + Vision
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
triton==3.2.0
faiss-cpu==1.10.0

# Face recognition
face-recognition==1.3.0
dlib==19.24.2

# Utilities
numpy==2.2.4
pillow==11.1.0
matplotlib==3.10.1
tqdm==4.67.1
regex==2024.11.6
ftfy==6.3.1
scikit-learn==1.4.2

# File watching
watchdog==4.0.0

# Other deps already present
python-dateutil==2.9.0.post0
six==1.17.0
typing_extensions==4.13.0
setuptools==78.1.0
